% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadm36_ALB_1.shp}
\alias{gadm36_ALB_1.shp}
\title{Shape file of Australia.}
\format{A .shp data file which is with 10 vectors and 1 list, we would focus on the list "geometry".
\describe{
  \item{geometry}{A list (deepth = 3) contains the latitudes and longitudes.}
}}
\source{
\url{https://gadm.org/download_country_v3.html}
}
\usage{
read_sf("data/gadm36_ALB_shp/gadm36_ALB_1.shp")
}
\description{
A dataset containing the shape details of Australia.
}
\keyword{datasets}
