% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team8.R
\name{helper}
\alias{helper}
\title{Trans a maxrix with longitudes and latitudes into a data frame.}
\usage{
helper(y)
}
\arguments{
\item{y}{a matrix with longitudes and latitudes.}
}
\value{
A data frame including long, lat, groups, and order.
}
\description{
Trans a maxrix with longitudes and latitudes into a data frame.
}
