% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team8.R
\name{team_8}
\alias{team_8}
\title{Plot a state shape in Australia.}
\usage{
team_8(file, tolerance = 0.1)
}
\arguments{
\item{file}{The name of the file includes the data (shp. format).}

\item{tolerance}{The tolerance level to thinen the origninal data. The default value is 0.1. The larger tolerance value, the thinner data.}
}
\value{
A list includes a vector of state name, and a data frame shows the polygon of longitudes and latitudes.
}
\description{
Plot a state shape in Australia.
}
\examples{
#Find the the path "fp" of your own shp file, then call team8(fp)
}
