% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_1.R
\name{team_1}
\alias{team_1}
\title{Plot a country shape with .shp file.}
\usage{
team_1(file, tolerance = 0.1)
}
\arguments{
\item{file}{The name of the file which the data are to be read from. The file should be in shp format.}

\item{tolerance}{The tolerance level to be used in thinning the origninal data. The default value is 0.1. The original data usually is too large to run. Larger tolerance value makes the data thinner.}
}
\value{
A list includes a vector of country name, a vector of the states name, and a data frame of the polygon longitudes and latitudes.
}
\description{
Plot a country shape with .shp file.
}
\examples{
#See details about this example data with ?gadm36_AUS_1.shp
fp <- paste(getwd(), "/data_raw/gadm36_AUS_shp/gadm36_AUS_1.shp", sep = "")
team_1(fp)
team_1(fp, 0.3)
}
